split up AtomicStatePlotter
