+++ ReadingOutputFromFLYonPIC_v2/reader/StateDistributionReader.py
@@ -18,14 +18,14 @@ from . import Reader
 
 @typeguard.typechecked
 class StateDistributionReader(Reader):
-    def read(self) -> tuple[npt.NDArray[np.float64], tuple[npt.NDArray[np.float64], npt.NDArray[np.uint64]], dict[str, int], tuple[np.float64]]:
+    def read(self) -> tuple[npt.NDArray[np.float64], tuple[npt.NDArray[np.float64], npt.NDArray[typing.Any]], dict[str, int], tuple[np.float64]]:
         """
         read in state distribution over time from source
 
         @returns
             data table ... accumulated weight of macro particles for atomic state and time step : np.ndarray((numberTimeSteps, numberAtomicStates)),
-            (list of times points, list of atomic state config numbers),
-            axisDict ... map of atomicState and timeStep physics axis to axis index in the data table : {"atomicState" : int, "timeStep": int},
-            (scaling factor)
+            (list of times points, list of states),
+            axisDict ... map of state and timeStep physics axis to axis index in the data table : {"atomicState/chargeState" : int, "timeStep": int},
+            (scaling factor,)
         """
         raise NotImplementedError("abstract interface only")
